+++ vm/src/vm/runners/cairo_runner.rs
@@ -188,6 +188,7 @@ impl CairoRunner {
             LayoutName::recursive_large_output => CairoLayout::recursive_large_output_instance(),
             LayoutName::recursive_with_poseidon => CairoLayout::recursive_with_poseidon(),
             LayoutName::all_cairo => CairoLayout::all_cairo_instance(),
+            LayoutName::all_cairo_stwo => CairoLayout::all_cairo_stwo_instance(),
             LayoutName::all_solidity => CairoLayout::all_solidity_instance(),
             LayoutName::dynamic => {
                 let params =
@@ -3401,6 +3402,29 @@ mod tests {
     fn run_empty() {
         let program = program!();
         let mut cairo_runner = cairo_runner!(&program, LayoutName::all_cairo, false, true);
+        assert_matches!(
+            cairo_runner.initialize(false),
+            Err(RunnerError::MissingMain)
+        );
+    }
+
+    #[test]
+    #[cfg_attr(target_arch = "wasm32", wasm_bindgen_test)]
+    fn run_empty_recursive_with_poseidon() {
+        let program = program!();
+        let mut cairo_runner =
+            cairo_runner!(&program, LayoutName::recursive_with_poseidon, false, true);
+        assert_matches!(
+            cairo_runner.initialize(false),
+            Err(RunnerError::MissingMain)
+        );
+    }
+
+    #[test]
+    #[cfg_attr(target_arch = "wasm32", wasm_bindgen_test)]
+    fn run_empty_all_cairo_stwo() {
+        let program = program!();
+        let mut cairo_runner = cairo_runner!(&program, LayoutName::all_cairo_stwo, false, true);
         assert_matches!(
             cairo_runner.initialize(false),
             Err(RunnerError::MissingMain)